requirements.txt
Flask==2.3.3
PyYAML==6.0.1
Jinja2==3.1.2
WTForms==3.0.1
Flask-WTF==1.1.1
ipaddress==1.0.23
